pyproject.toml
[project]
name = "genai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "easyocr>=1.7.2",
    "faiss-cpu>=1.13.2",
    "huggingface-hub>=1.7.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "nltk>=3.9.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
